Created IPAddressMailer.py
The file can be used for the following purpose:
    1. mailing the current IP address to some predefined email id 
    2. Just for sending mail using python script.

Edit the uname and password with your username and password. destUname is the email id where you want to send the mail.

For Ubuntu, create an entry using StartUp Applications for running it each time the computer turns on to get your IPAddress if you are using DHCP and remotely want to access the host.